requirements.txt
alabaster==0.7.12
asgiref==3.6.0
attrs==19.1.0
Babel==2.7.0
bcrypt==4.0.1
capstone==5.0.0rc2
certifi==2019.6.16
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
cipher-solver==1.0.1
cloudinary==1.32.0
colorama==0.4.1
colored-traceback==0.3.0
crypto==1.4.1
cryptography==38
cycler==0.11.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.15.2
filelock==3.9.0
fonttools==4.38.0
gunicorn==20.1.0
idna==2.8
imagesize==1.1.0
intervaltree==3.1.0
Jinja2==2.10.1
joblib==1.2.0
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==1.1.1
mpmath==1.2.1
Naked==0.1.32
packaging==19.1
paramiko==3.0.0
Pillow==9.4.0
platformdirs==3.0.0
plumbum==1.8.1
psutil==5.9.4
psycopg2==2.9.5
pwn==1.0
pwntools==4.9.0
pycparser==2.21
pycryptodomex==3.17
pyelftools==0.29
Pygments==2.4.2
PyJWT==1.7.1
PyNaCl==1.5.0
pyparsing==2.4.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2019.1
PyYAML==6.0
pyzipper==0.3.6
requests==2.25.1
ROPGadget==7.2
rpyc==5.3.1
shellescape==3.8.1
six==1.12.0
snowballstemmer==1.9.0
sortedcontainers==2.4.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.1
sympy==1.11.1
threadpoolctl==3.1.0
timeout-decorator==0.5.0
typing_extensions==3.10.0.0
tzdata==2023.3
unicorn==2.0.1.post1
unrar==0.4
urllib3==1.26.5
virtualenv==20.19.0
whitenoise==6.4.0
python-dotenv==1.0.0
boto3==1.34.21
